restructure sagemaker test run functions
